backend/requirements.txt
# Django and REST Framework
Django>=5.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Configuration
python-decouple>=3.8,<4.0

# Production Server
gunicorn>=21.0,<24.0

# Static Files (for production)
whitenoise>=6.0,<7.0
